CF2-016: Using the request pattern to dashboard requests
